+++ inginious/frontend/pages/course_admin/task_list.py
@@ -10,6 +10,7 @@ from collections import OrderedDict
 from natsort import natsorted
 
 from inginious.frontend.pages.course_admin.utils import INGIniousAdminPage
+from inginious.frontend.util import change_access_structure, dict_data_str_to_datetimes
 
 class CourseTaskListPage(INGIniousAdminPage):
     """ List informations about all tasks """
@@ -37,6 +38,9 @@ class CourseTaskListPage(INGIniousAdminPage):
             task_dispenser = course.get_task_dispenser()
             try:
                 data = task_dispenser.import_legacy_tasks()
+                for taskid, task in data["config"].items():
+                    task["accessibility"] = dict_data_str_to_datetimes(
+                            change_access_structure(task["accessibility"], True))
                 self.update_dispenser(course, data)
             except Exception as e:
                 errors.append(_("Something wrong happened: ") + str(e))
