+++ inginious/frontend/pages/taskset_admin/template.py
@@ -10,7 +10,7 @@ from collections import OrderedDict
 from natsort import natsorted
 
 from inginious.frontend.pages.taskset_admin.utils import INGIniousAdminPage
-
+from inginious.frontend.util import change_access_structure, dict_data_str_to_datetimes
 
 
 class TasksetTemplatePage(INGIniousAdminPage):
@@ -39,6 +39,9 @@ class TasksetTemplatePage(INGIniousAdminPage):
             task_dispenser = taskset.get_task_dispenser()
             try:
                 data = task_dispenser.import_legacy_tasks()
+                for taskid, task in data["config"].items():
+                    task["accessibility"] = dict_data_str_to_datetimes(
+                        change_access_structure(task["accessibility"], True))
                 self.update_dispenser(taskset, data)
             except Exception as e:
                 errors.append(_("Something wrong happened: ") + str(e))
